UCP/PROTO: Move rndv files to separate sub-folder
